bugfix:    - Moved GoalsActivity to package presentation.view    - Removed UI package    - Removed duplicated package util
