+++ app/src/main/java/com/app/fitspace/presentation/view/GoalsActivity.kt
@@ -1,4 +1,4 @@
-package com.app.fitspace.ui
+package com.app.fitspace.presentation.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
