--- app/src/main/java/com/app/fitspace/util/DateUtils.kt
@@ -1,4 +1,0 @@
-package com.app.fitspace.util
-
-class DateUtils {
-}
